feat: Add business location and zone features
This commit introduces location data (address, latitude, longitude) to businesses, enabling geographical search and mapping. It also defines `Zone` types and a list of predefined zones, along with updates to the `ZonesPage` to display businesses filtered by location and calculated distance. A new `MapPinIcon` component is added for visual representation.